+++ addon/index.js
@@ -18,6 +18,7 @@ import OfflineModel from './models/offline-model';
 import OfflineModelMixin from './mixins/offline-model';
 import BaseStore from './stores/base-store';
 import LocalStore from './stores/local-store';
+import OnlineStore from './stores/online-store';
 import OfflineSerializer from './serializers/offline';
 import OfflineAdapter from './adapters/offline';
 import OfflineGlobalsService from './services/offline-globals';
@@ -39,15 +40,19 @@ import JsAdapter from './query/js-adapter';
 import QueryOdataAdapter from './query/odata-adapter';
 import OrderByClause from './query/order-by-clause';
 import QueryObject from './query/query-object';
-import { BasePredicate } from './query/predicate';
-import { SimplePredicate } from './query/predicate';
-import { DatePredicate } from './query/predicate';
-import { ComplexPredicate }from './query/predicate';
-import { StringPredicate }from './query/predicate';
-import { DetailPredicate } from './query/predicate';
-import { createPredicate } from './query/predicate';
-import { GeographyPredicate } from './query/predicate';
-import { GeometryPredicate } from './query/predicate';
+import {
+  BasePredicate,
+  SimplePredicate,
+  DatePredicate,
+  ComplexPredicate,
+  StringPredicate,
+  DetailPredicate,
+  GeographyPredicate,
+  GeometryPredicate,
+  NotPredicate,
+  IsOfPredicate,
+  createPredicate,
+} from './query/predicate';
 
 import UserService from './services/user';
 import DexieService from './services/dexie';
@@ -85,6 +90,7 @@ Projection.create = create;
 Projection.attr = attr;
 Projection.belongsTo = belongsTo;
 Projection.hasMany = hasMany;
+Projection.OnlineStore = OnlineStore;
 
 /**
   This namespace contains base adapter classes for different kind of backends.
@@ -134,9 +140,11 @@ Query.DatePredicate = DatePredicate;
 Query.ComplexPredicate = ComplexPredicate;
 Query.StringPredicate = StringPredicate;
 Query.DetailPredicate = DetailPredicate;
-Query.createPredicate = createPredicate;
 Query.GeographyPredicate = GeographyPredicate;
 Query.GeometryPredicate = GeometryPredicate;
+Query.NotPredicate = NotPredicate;
+Query.IsOfPredicate = IsOfPredicate;
+Query.createPredicate = createPredicate;
 
 /**
   This namespace contains helper claseses.