+++ addon/query/indexeddb-adapter.js
@@ -616,7 +616,7 @@ function updateWhereClause(store, table, query) {
         break;
 
       case 'date':
-        value = getSerializedDateValue.call(store, predicate.value);
+        value = getSerializedDateValue.call(store, predicate.value, predicate.timeless);
         break;
 
       default:
@@ -631,21 +631,36 @@ function updateWhereClause(store, table, query) {
       return table.filter(jsAdapter.getAttributeFilterFunction(predicate));
     }
 
+    let moment;
+    let nextValue;
+    if (predicate.timeless) {
+      moment = Ember.getOwner(store).lookup('service:moment');
+      nextValue = moment.moment(value, 'YYYY-MM-DD').add(1, 'd').format('YYYY-MM-DD');
+    }
+
     switch (predicate.operator) {
       case FilterOperator.Eq:
-        return table.where(predicate.attributePath).equals(value);
+        return predicate.timeless ?
+          table.where(predicate.attributePath).between(value, nextValue, false) :
+          table.where(predicate.attributePath).equals(value);
 
       case FilterOperator.Neq:
-        return table.where(predicate.attributePath).notEqual(value);
+        return predicate.timeless ?
+          table.where(predicate.attributePath).below(value).or(predicate.attributePath).aboveOrEqual(nextValue) :
+          table.where(predicate.attributePath).notEqual(value);
 
       case FilterOperator.Le:
         return table.where(predicate.attributePath).below(value);
 
       case FilterOperator.Leq:
-        return table.where(predicate.attributePath).belowOrEqual(value);
+        return predicate.timeless ?
+          table.where(predicate.attributePath).below(nextValue) :
+          table.where(predicate.attributePath).belowOrEqual(value);
 
       case FilterOperator.Ge:
-        return table.where(predicate.attributePath).above(value);
+        return predicate.timeless ?
+          table.where(predicate.attributePath).aboveOrEqual(nextValue) :
+          table.where(predicate.attributePath).above(value);
 
       case FilterOperator.Geq:
         return table.where(predicate.attributePath).aboveOrEqual(value);