+++ addon/stores/online-store.js
@@ -1,0 +1,32 @@
+/**
+  @module ember-flexberry-data
+*/
+
+import Ember from 'ember';
+import DS from 'ember-data';
+
+/**
+  Store that used in online mode by default.
+
+  @class OnlineStore
+  @namespace OData
+  @extends <a href="http://emberjs.com/api/data/classes/DS.Store.html">DS.Store</a>
+  @private
+*/
+export default DS.Store.extend({
+
+  /**
+    Delete all record from the current store.
+    @method deleteRecord
+    @param {String} modelName modelName
+    @param {Object} filter filter
+  */
+  deleteAllRecords: function(modelName, filter) {
+    let adapter = this.adapterFor(modelName);
+    if (Ember.isNone(adapter.deleteAllRecords)) {
+      Ember.assert('Method \'deleteAllRecords\' is missing');
+    }
+
+    return adapter.deleteAllRecords(adapter.store, modelName, filter);
+  }
+});